+++ src/RcsbFvSequence/SequenceViews/RcsbView/Components/HelpLinkComponent.tsx
@@ -1,0 +1,33 @@
+import * as React from "react";
+import {RcsbFvDOMConstants} from "../../../../RcsbFvConstants/RcsbFvConstants";
+
+interface HelpLinkInterface {
+    unmount:(flag:boolean,callback:()=>void)=>void;
+    helpHref:string;
+}
+
+export class HelpLinkComponent extends React.Component<HelpLinkInterface> {
+    render(): JSX.Element {
+        return(
+            <div style={{marginTop:10}}>
+                <div>
+                    <div id={RcsbFvDOMConstants.SELECT_BUTTON_PFV_ID} style={{display:"inline-block"}}/>
+                </div>
+                <div style={{position:"absolute", top:5, right:5}} >
+                    <a
+                        style={{textDecoration:"none", color:"#337ab7", cursor:"pointer", marginRight:15}}
+                        target={"_blank"}
+                        href={this.props.helpHref}
+                    >
+                        Help
+                    </a>
+                    <a style={{textDecoration:"none", color: "#337ab7", cursor:"pointer"}} onClick={()=>{this.props.unmount(true, ()=>{
+                        window.history.back();
+                    })}}>
+                        Back
+                    </a>
+                </div>
+            </div>
+        );
+    }
+}