+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 [Semantic Versioning](https://semver.org/)
 
+## [2.2.1] - 2022-11-03
+### Improvement
+- New attribute `RcsbViewInterface.additionalContent` to define the `additionalContent` React component
+
 ## [2.2.0] - 2022-11-03
 ### New Features
 - New UniProt MSA 1D3D view
@@ -9,6 +13,14 @@
   - `SequenceIdentityPfvManagerFactory` builds Sequence Identity MSA PFV
   - `MsaCallbackManagerFactory` MSA 1D callbacks
   - `MsaBehaviourObserver` MSA 3D callbacks
+
+## [2.1.1] - 2022-10-17
+### Dependency update
+- rcsb-saguaro-app v4.4.13
+- rcsb-saguaro v2.2.16
+
+### Minor display update
+- 1D PFV on-change resets 3D display
 
 ## [2.1.0] - 2022-09-02
 ### Major refactoring